** ReferenceSpace:      -- Minimal number of reference space (at least for tet, tri and quad)      -- Automatic generation of these reference space      -- New interface      -- Modification of TriNodeBasis accordingly
** Need to compute Dof permutations automaticaly

** All Basis are commented (exept for TriNodeBasis)

** WARNING: Memory leak somewhere

+++ FunctionSpace/TriNodeBasis.h
@@ -20,7 +20,7 @@ class TriNodeBasis: public BasisHierarchical0From{
   //! @param order The order of the Basis
   //!
   //! Returns a new Node-Basis for Triangles of the given order
-  TriNodeBasis(unsigned int order);
+  TriNodeBasis(size_t order);
 
   //! Deletes this Basis
   //!
